@@ -1,17 +1,18 @@
 name: genius_api_unofficial
 description: This library provides you with a nice interface to Genius API.
-version: 1.0.0
+version: 2.0.0
 homepage: https://github.com/nt4f04uNd/genius_api_unofficial
 repository: https://github.com/nt4f04uNd/genius_api_unofficial
 issue_tracker: https://github.com/nt4f04uNd/genius_api_unofficial/issues
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  meta: ^1.3.0
+  meta: ^1.7.0
   http: ^0.13.3
-  
+  enum_to_string: ^2.0.1
+
 dev_dependencies:
-  pedantic: ^1.11.0
+  flutter_lints: ^1.0.4
   test: ^1.17.4
